:hammer: Move file observer from issues to reloadable_json module
